Merge to main

+++ timexer/data_provider/data_loader.py
@@ -293,10 +293,10 @@ class Dataset_Custom(Dataset):
         r_begin = s_end - self.label_len
         r_end = r_begin + self.label_len + self.pred_len
 
-        seq_x = self.data_x[s_begin:s_end]
-        seq_y = self.data_y[r_begin:r_end]
-        seq_x_mark = self.data_stamp[s_begin:s_end]
-        seq_y_mark = self.data_stamp[r_begin:r_end]
+        seq_x = self.data_x[s_begin:s_end].astype('float32')
+        seq_y = self.data_y[r_begin:r_end].astype('float32')
+        seq_x_mark = self.data_stamp[s_begin:s_end].astype('float32')
+        seq_y_mark = self.data_stamp[r_begin:r_end].astype('float32')
 
         return seq_x, seq_y, seq_x_mark, seq_y_mark
